buy stock works

+++ app/models/user.rb
@@ -17,6 +17,6 @@ class User < ActiveRecord::Base
   has_secure_password
 
   def positions_total
-    stocks.map(&:position).reduce(:+).to_i
+    stocks.map(&:position).reduce(:+).to_f
   end
 end
